+++ libc/src/ctype/isalpha.cpp
@@ -9,12 +9,14 @@
 #include "src/ctype/isalpha.h"
 
 #include "src/__support/common.h"
-#include "src/ctype/ctype_utils.h"
+#include "src/__support/ctype_utils.h"
 
 namespace __llvm_libc {
 
 // TODO: Currently restricted to default locale.
 // These should be extended using locale information.
-LLVM_LIBC_FUNCTION(int, isalpha, (int c)) { return internal::isalpha(c); }
+LLVM_LIBC_FUNCTION(int, isalpha, (int c)) {
+  return static_cast<int>(internal::isalpha(static_cast<unsigned>(c)));
+}
 
 } // namespace __llvm_libc
